requirements.txt
ttkbootstrap
pillow
pyperclip
pygame
edge-tts